Switch to virtual method dispatching
Sad moment but the nested templates were becoming too much. Makes it
really hard to extend things. Pretty minimal performance benefit as
well, except in some edge cases like with the Automaton snail skipping a
large amount of computation through extensive inlining.

Even though it's slightly slower, this results in a *substantially* smaller bundle size:

Before:
- snail_lattice.js            50.4 KB
- snail_lattice_bg.wasm       304.7 KB

After:
- snail_lattice.js            11.9 KB
- snail_lattice_bg.wasm       179.9 KB

Or a roughly 46% decrease on the WASM side.

+++ snail-lattice/src/direction.rs
@@ -17,8 +17,8 @@ impl Direction {
         }
     }
 
-    pub fn to_wall(self) -> usize {
-        1 << (3 - self as usize)
+    pub fn to_wall(self) -> u8 {
+        1 << (3 - self as u8)
     }
 
     pub fn rotate(self) -> Direction {
